move stats decoder to a service
